Add files via upload
This repository contains the implementation of a machine learning model to predict diamond prices based on various features such as carat weight, cut, color, clarity, etc. The goal of this project is to develop an accurate predictive model that can assist in estimating the price of diamonds based on their characteristics.